check before delete production

+++ src/page/AdminProduct.js
@@ -59,8 +59,11 @@ const AdminProduct = () => {
 
   //아이템 삭제하기
   const deleteItem = (id) => {
-    dispatch(productActions.deleteProduct(id));
-    dispatch(productActions.getProductList(searchQuery));
+    const isConfirmed = window.confirm("해당 상품을 삭제하시겠습니까?");
+    if (isConfirmed) {
+      dispatch(productActions.deleteProduct(id));
+      dispatch(productActions.getProductList(searchQuery));
+    }
   };
 
   const openEditForm = (product) => {
